feat(workspace): flag the dock chat stage, refine focus ring, move new-chat to the left rail
The dockview-backed stage is now opt-in behind a chatPaneEngine feature
flag ("flex" | "dock"): explicit prop on WorkspaceAgentFront/ChatLayout,
else the boring-workspace:chat-pane-engine localStorage override, else
the flex row default. The flex stage is restored as the default engine;
both implementations share the pane contract (focus, flash, close
semantics, aria/test attributes).

Dock mode loses the tab look per design feedback: each pane gets a flat
full-width header — drag grip + title + close, no tab pill — and drops
onto tab strips or pane centers are vetoed via onWillShowOverlay, so a
drag can only split at pane edges (panes split, they never stack).

The active-pane treatment is now a soft rounded inset ring that fades
in/out (200ms) instead of a hard popping border, shared by both engines.
The floating "New chat" button moves from the stage's right edge to the
left-edge control stack directly above the Sessions toggle, creating
next to the active pane.
