+++ packages/workspace/src/front/layout/ChatPaneStageDock.tsx
@@ -1,0 +1,347 @@
+import {
+  createContext,
+  useCallback,
+  useContext,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from "react"
+import {
+  DockviewReact,
+  type DockviewApi,
+  type DockviewReadyEvent,
+  type IDockviewPanelHeaderProps,
+  type IDockviewPanelProps,
+} from "dockview-react"
+import "dockview-react/dist/styles/dockview.css"
+import "../dock/dockview-overrides.css"
+import "./chat-pane-stage.css"
+import { GripVertical, X } from "lucide-react"
+import { IconButton } from "@hachej/boring-ui-kit"
+import { cn } from "../lib/utils"
+import { PaneFocusRing, paneTitle, type ChatPaneDescriptor, type ChatPaneStageProps } from "./ChatPaneStage"
+
+type ChatPaneStageDockProps = Omit<ChatPaneStageProps, "engine">
+
+const CHAT_PANE_COMPONENT = "chat-pane"
+const PANE_MIN_WIDTH = 280
+const PERSIST_DEBOUNCE_MS = 300
+
+interface StageContextValue {
+  panes: ChatPaneDescriptor[]
+  activePaneId: string | null
+  flashPaneId: string | null
+  renderPane: ChatPaneStageProps["renderPane"]
+  onActivePaneChange?: (id: string) => void
+  onClosePane?: (id: string) => void
+}
+
+const StageContext = createContext<StageContextValue | null>(null)
+
+function layoutStorageKey(storageKey: string): string {
+  return `${storageKey}:chatPaneLayout`
+}
+
+interface SerializedDockLayout {
+  grid?: unknown
+  panels?: Record<string, unknown>
+}
+
+function readStoredLayout(storageKey: string, paneIds: string[]): SerializedDockLayout | null {
+  try {
+    const raw = globalThis.localStorage?.getItem(layoutStorageKey(storageKey))
+    if (!raw) return null
+    const parsed = JSON.parse(raw) as SerializedDockLayout
+    const storedIds = Object.keys(parsed.panels ?? {})
+    if (storedIds.length !== paneIds.length) return null
+    const wanted = new Set(paneIds)
+    if (!storedIds.every((id) => wanted.has(id))) return null
+    return parsed
+  } catch {
+    return null
+  }
+}
+
+function writeStoredLayout(storageKey: string, layout: unknown): void {
+  try {
+    globalThis.localStorage?.setItem(layoutStorageKey(storageKey), JSON.stringify(layout))
+  } catch {
+    // Best-effort persistence only.
+  }
+}
+
+function addChatPanel(
+  api: DockviewApi,
+  pane: ChatPaneDescriptor,
+  position: Parameters<DockviewApi["addPanel"]>[0]["position"],
+): void {
+  const panel = api.addPanel({
+    id: pane.id,
+    component: CHAT_PANE_COMPONENT,
+    title: paneTitle(pane),
+    params: { paneId: pane.id },
+    position,
+  })
+  panel.group?.api.setConstraints({ minimumWidth: PANE_MIN_WIDTH })
+}
+
+/**
+ * Project the authoritative pane list onto the dockview layout: remove
+ * panels whose session pane closed, add new ones next to their list
+ * neighbour, keep titles fresh, and align the active panel. Geometry that
+ * the user arranged (splits, sizes) is dockview's to keep.
+ */
+function syncPanesToDock(
+  api: DockviewApi,
+  panes: ChatPaneDescriptor[],
+  activePaneId: string | null,
+): void {
+  const wanted = new Map(panes.map((pane) => [pane.id, pane]))
+  for (const panel of [...api.panels]) {
+    if (!wanted.has(panel.id)) api.removePanel(panel)
+  }
+  panes.forEach((pane, index) => {
+    if (api.getPanel(pane.id)) return
+    const before = index > 0 ? api.getPanel(panes[index - 1].id) : undefined
+    const after = !before && index + 1 < panes.length ? api.getPanel(panes[index + 1].id) : undefined
+    addChatPanel(
+      api,
+      pane,
+      before
+        ? { referencePanel: before, direction: "right" }
+        : after
+          ? { referencePanel: after, direction: "left" }
+          : undefined,
+    )
+  })
+  for (const pane of panes) {
+    const panel = api.getPanel(pane.id)
+    if (panel && panel.title !== paneTitle(pane)) panel.api.setTitle(paneTitle(pane))
+  }
+  if (activePaneId) {
+    const panel = api.getPanel(activePaneId)
+    if (panel && api.activePanel?.id !== activePaneId) panel.api.setActive()
+  }
+}
+
+/** Dockview-backed chat stage: drag pane headers to split in any direction. */
+export function ChatPaneStageDock({
+  panes,
+  activePaneId,
+  renderPane,
+  onActivePaneChange,
+  onClosePane,
+  flashPaneId,
+  storageKey,
+}: ChatPaneStageDockProps) {
+  const apiRef = useRef<DockviewApi | null>(null)
+  // True while this component mutates dockview itself; dockview activation
+  // events fired during programmatic sync must not echo back to the parent.
+  const syncingRef = useRef(false)
+  const disposeRef = useRef<(() => void) | null>(null)
+
+  const latestRef = useRef({ panes, activePaneId: activePaneId ?? null, onActivePaneChange, storageKey })
+  latestRef.current = { panes, activePaneId: activePaneId ?? null, onActivePaneChange, storageKey }
+
+  const resolvedActiveId = activePaneId ?? panes[0]?.id ?? null
+
+  const contextValue = useMemo<StageContextValue>(() => ({
+    panes,
+    activePaneId: resolvedActiveId,
+    flashPaneId: flashPaneId ?? null,
+    renderPane,
+    onActivePaneChange,
+    onClosePane: panes.length > 1 ? onClosePane : undefined,
+  }), [panes, resolvedActiveId, flashPaneId, renderPane, onActivePaneChange, onClosePane])
+
+  const handleReady = useCallback((event: DockviewReadyEvent) => {
+    const api = event.api
+    apiRef.current = api
+    const { panes: currentPanes, activePaneId: currentActive, storageKey: currentKey } = latestRef.current
+
+    syncingRef.current = true
+    try {
+      const stored = currentKey ? readStoredLayout(currentKey, currentPanes.map((pane) => pane.id)) : null
+      if (stored) {
+        try {
+          api.fromJSON(stored as Parameters<DockviewApi["fromJSON"]>[0])
+        } catch {
+          // A stale/incompatible layout must never block the chat stage.
+        }
+      }
+      syncPanesToDock(api, currentPanes, currentActive)
+    } finally {
+      syncingRef.current = false
+    }
+
+    const activeDisposable = api.onDidActivePanelChange((panel) => {
+      if (syncingRef.current) return
+      const id = panel?.id
+      if (id && id !== latestRef.current.activePaneId) {
+        latestRef.current.onActivePaneChange?.(id)
+      }
+    })
+
+    // Panes split, they don't stack: veto tab-strip and center drops so a
+    // drag can only dock to pane edges.
+    const overlayDisposable = api.onWillShowOverlay((overlayEvent) => {
+      if (overlayEvent.kind === "tab" || overlayEvent.kind === "header_space") {
+        overlayEvent.preventDefault()
+        return
+      }
+      if (overlayEvent.kind === "content" && overlayEvent.position === "center") {
+        overlayEvent.preventDefault()
+      }
+    })
+
+    let persistTimer: ReturnType<typeof setTimeout> | null = null
+    const layoutDisposable = api.onDidLayoutChange(() => {
+      const key = latestRef.current.storageKey
+      if (!key) return
+      if (persistTimer) clearTimeout(persistTimer)
+      persistTimer = setTimeout(() => writeStoredLayout(key, api.toJSON()), PERSIST_DEBOUNCE_MS)
+    })
+
+    disposeRef.current = () => {
+      if (persistTimer) clearTimeout(persistTimer)
+      activeDisposable.dispose()
+      overlayDisposable.dispose()
+      layoutDisposable.dispose()
+    }
+  }, [])
+
+  useEffect(() => () => disposeRef.current?.(), [])
+
+  useEffect(() => {
+    const api = apiRef.current
+    if (!api) return
+    syncingRef.current = true
+    try {
+      syncPanesToDock(api, panes, resolvedActiveId)
+    } finally {
+      syncingRef.current = false
+    }
+  }, [panes, resolvedActiveId])
+
+  if (panes.length === 0) return null
+
+  return (
+    <StageContext.Provider value={contextValue}>
+      <div
+        data-boring-workspace-part="chat-pane-stage"
+        className="relative h-full min-h-0 w-full bg-background"
+      >
+        <DockviewReact
+          className="dv-shell dv-chat-stage h-full"
+          components={STAGE_COMPONENTS}
+          defaultTabComponent={ChatPaneHeader as React.FunctionComponent<IDockviewPanelHeaderProps>}
+          onReady={handleReady}
+        />
+      </div>
+    </StageContext.Provider>
+  )
+}
+
+function useStage(): StageContextValue {
+  const value = useContext(StageContext)
+  if (!value) throw new Error("Chat pane components must render inside ChatPaneStageDock")
+  return value
+}
+
+function ChatPanePanel(props: IDockviewPanelProps) {
+  const stage = useStage()
+  const paneId = typeof (props.params as { paneId?: unknown })?.paneId === "string"
+    ? (props.params as { paneId: string }).paneId
+    : props.api.id
+  const pane = stage.panes.find((candidate) => candidate.id === paneId)
+  if (!pane) return null
+
+  const active = paneId === stage.activePaneId
+  const multiPane = stage.panes.length > 1
+  const flash = paneId === stage.flashPaneId
+  return (
+    <div
+      data-boring-workspace-part="chat-pane"
+      data-boring-state={active ? "active" : "inactive"}
+      aria-label={`Chat session ${paneTitle(pane)}`}
+      className="relative flex h-full min-h-0 w-full flex-col overflow-hidden bg-background"
+      onMouseDown={(event) => {
+        const target = event.target instanceof HTMLElement ? event.target : null
+        if (target?.closest('[data-boring-workspace-part="chat-pane-control"]')) return
+        stage.onActivePaneChange?.(paneId)
+      }}
+      onFocusCapture={(event) => {
+        const target = event.target instanceof HTMLElement ? event.target : null
+        if (target?.closest('[data-boring-workspace-part="chat-pane-control"]')) return
+        stage.onActivePaneChange?.(paneId)
+      }}
+    >
+      <PaneFocusRing active={multiPane && active} flash={flash} />
+      <div className="min-h-0 flex-1 overflow-hidden">
+        {stage.renderPane(pane)}
+      </div>
+    </div>
+  )
+}
+
+const STAGE_COMPONENTS: Record<string, React.FunctionComponent<IDockviewPanelProps>> = {
+  [CHAT_PANE_COMPONENT]: ChatPanePanel,
+}
+
+/**
+ * Flat pane header — not a tab. The whole bar is dockview's drag handle;
+ * the grip is the visual affordance for it, the X closes the view.
+ */
+function ChatPaneHeader(props: IDockviewPanelHeaderProps) {
+  const stage = useStage()
+  const { api } = props
+  const [title, setTitle] = useState(api.title ?? api.id)
+
+  useEffect(() => {
+    const sync = () => setTitle(api.title ?? api.id)
+    sync()
+    const sub = api.onDidTitleChange?.(sync)
+    return () => sub?.dispose?.()
+  }, [api])
+
+  const canClose = Boolean(stage.onClosePane)
+  return (
+    <div
+      className="group flex h-full w-full min-w-0 cursor-grab select-none items-center gap-1.5 px-2 text-[12px] font-medium leading-none tracking-tight active:cursor-grabbing"
+      title={title}
+    >
+      <GripVertical
+        aria-hidden="true"
+        className="h-3.5 w-3.5 shrink-0 text-muted-foreground/45 transition-colors group-hover:text-muted-foreground"
+        strokeWidth={1.75}
+      />
+      <span className="min-w-0 flex-1 overflow-hidden text-ellipsis whitespace-nowrap text-foreground/70">
+        {title}
+      </span>
+      {canClose ? (
+        <IconButton
+          type="button"
+          variant="ghost"
+          size="icon-xs"
+          data-boring-workspace-part="chat-pane-control"
+          className="h-5 w-5 shrink-0 text-muted-foreground/80 opacity-0 focus-visible:opacity-100 group-hover:opacity-100 [.dv-active-tab_&]:opacity-55 [.dv-active-tab_&]:hover:opacity-100"
+          // Dockview activates a panel from a NATIVE pointerdown listener on
+          // the tab wrapper (an ancestor of this button). React's capture
+          // handlers run at root-capture, before that bubble listener — stop
+          // the native event there so closing a pane never activates it.
+          onPointerDownCapture={(event) => event.nativeEvent.stopPropagation()}
+          onMouseDownCapture={(event) => event.nativeEvent.stopPropagation()}
+          onClick={(event) => {
+            event.preventDefault()
+            event.stopPropagation()
+            stage.onClosePane?.(api.id)
+          }}
+          aria-label={`Close ${title} pane`}
+        >
+          <X className="h-3 w-3" strokeWidth={2.25} />
+        </IconButton>
+      ) : null}
+    </div>
+  )
+}
